improved vacation service return type fro graphql

+++ TimeTracker.Server.Business/Services/VacationService.cs
@@ -1,6 +1,5 @@
 ﻿using AutoMapper;
 using GraphQL;
-using Microsoft.AspNetCore.Http;
 using TimeTracker.Server.Business.Abstractions;
 using TimeTracker.Server.Business.Models.Vacation;
 using TimeTracker.Server.Data.Abstractions;
@@ -8,6 +7,7 @@ using TimeTracker.Server.Shared.Exceptions;
 using TimeTracker.Server.Shared.Helpers;
 using TimeTracker.Server.Shared;
 using TimeTracker.Server.Data.Models.Vacation;
+using TimeTracker.Server.Business.Models.User;
 
 namespace TimeTracker.Server.Business.Services;
 
@@ -17,76 +17,127 @@ public class VacationService : IVacationService
 
     private readonly IVacationInfoRepository _vacationInfoRepository;
     private readonly IVacationRepository _vacationRepository;
+    private readonly IUserRepository _userRepository;
 
     private readonly IUserService _userService;
 
-    public VacationService(IMapper mapper, IVacationInfoRepository vacationInfoRepository, IVacationRepository vacationRepository, IUserService userService)
+    public VacationService(IMapper mapper, IVacationInfoRepository vacationInfoRepository, IVacationRepository vacationRepository,
+        IUserRepository userRepository, IUserService userService)
     {
         _mapper = mapper;
         _vacationInfoRepository = vacationInfoRepository;
         _vacationRepository = vacationRepository;
+        _userRepository = userRepository;
         _userService = userService;
     }
 
-    public async Task<VacationBusinessResponse> GetVacationByIdAsync(Guid id)
+    public async Task<IEnumerable<VacationWithUserBusinessResponse>> GetVacationsByUserIdAsync(Guid userId, bool? onlyApproved, bool orderByDesc)
     {
-        var vacationDataResponse = await _vacationRepository.GetVacationByIdAsync(id);
-        if (vacationDataResponse is null)
-        {
-            throw new ExecutionError("Vacation not found")
-            {
-                Code = GraphQLCustomErrorCodesEnum.VACATION_NOT_FOUND.ToString()
-            };
-        }
-
         var curUser = await _userService.GetCurrentUserFromClaimsAsync();
-        if (curUser is null)
+        if (!PermissionHelper.HasPermit(curUser.Permissions, PermissionsEnum.GetVacation.ToString()))
         {
-            throw new ExecutionError("User not found")
+            if (curUser.Id != userId)
             {
-                Code = GraphQLCustomErrorCodesEnum.USER_NOT_FOUND.ToString()
-            };
-        }
-        else if (!PermissionHelper.HasPermit(curUser.Permissions, PermissionsEnum.GetVacation.ToString()))
-        {
-            if (curUser.Id != vacationDataResponse.UserId)
-            {
-                throw new ExecutionError("User do not has permission to read this vacation")
+                throw new ExecutionError("User do not has permission to read vacations")
                 {
                     Code = GraphQLCustomErrorCodesEnum.NO_PERMISSION.ToString()
                 };
             }
         }
 
-        var vacationBusinessResponse = _mapper.Map<VacationBusinessResponse>(vacationDataResponse);
-        return vacationBusinessResponse;
-    }
-
-    public async Task<IEnumerable<VacationBusinessResponse>> GetVacationsByUserIdAsync(Guid userId, bool? onlyApproved, bool orderByDesc)
-    {
         var vacationsDataResponse = await _vacationRepository.GetVacationsByUserIdAsync(userId, onlyApproved, orderByDesc);
         var vacationsBusinessResponse = _mapper.Map<IEnumerable<VacationBusinessResponse>>(vacationsDataResponse);
-        return vacationsBusinessResponse;
+
+        var userDataResponse = await _userRepository.GetUserByIdAsync(userId);
+        var userBusinessResponse = _mapper.Map<UserBusinessResponse>(userDataResponse);
+
+        var approverDict = new Dictionary<Guid, UserBusinessResponse?>();
+
+        var vacationWithUserBusinessResponses = new List<VacationWithUserBusinessResponse>();
+        foreach (var vacation in vacationsBusinessResponse)
+        {
+            UserBusinessResponse? approver;
+            if (vacation.ApproverId is null)
+            {
+                approver = null;
+            }
+            else
+            {
+                if (approverDict.ContainsKey((Guid)vacation.ApproverId))
+                {
+                    approver = approverDict[(Guid)vacation.ApproverId];
+                }
+                else
+                {
+                    var approverDataResponse = await _userRepository.GetUserByIdAsync((Guid)vacation.ApproverId);
+                    approver = _mapper.Map<UserBusinessResponse>(approverDataResponse);
+
+                    approverDict.Add(approver.Id, approver);
+                }
+            }
+
+            vacationWithUserBusinessResponses.Add(new VacationWithUserBusinessResponse()
+            {
+                Vacation = vacation,
+                User = userBusinessResponse,
+                Aprover = approver
+            });
+        }
+
+        return vacationWithUserBusinessResponses;
     }
 
-    public async Task<IEnumerable<VacationBusinessResponse>> GetVacationRequestsAsync()
+    public async Task<IEnumerable<VacationWithUserBusinessResponse>> GetVacationRequestsAsync(bool getNotStarted)
     {
-        var vacationsDataResponse = await _vacationRepository.GetVacationRequestsAsync();
+        var vacationsDataResponse = getNotStarted ?
+            await _vacationRepository.GetNotStartedUpdatedVacationsAsync() :
+            await _vacationRepository.GetVacationRequestsAsync();
+
         var vacationsBusinessResponse = _mapper.Map<IEnumerable<VacationBusinessResponse>>(vacationsDataResponse);
-        return vacationsBusinessResponse;
+
+        var approverDict = new Dictionary<Guid, UserBusinessResponse?>();
+
+        var vacationWithUserBusinessResponses = new List<VacationWithUserBusinessResponse>();
+        foreach (var vacation in vacationsBusinessResponse)
+        {
+            var userDataResponse = await _userRepository.GetUserByIdAsync(vacation.UserId);
+            var user = _mapper.Map<UserBusinessResponse>(userDataResponse);
+
+            UserBusinessResponse? approver;
+            if (vacation.ApproverId is null)
+            {
+                approver = null;
+            }
+            else
+            {
+                if (approverDict.ContainsKey((Guid)vacation.ApproverId))
+                {
+                    approver = approverDict[(Guid)vacation.ApproverId];
+                }
+                else
+                {
+                    var approverDataResponse = await _userRepository.GetUserByIdAsync((Guid)vacation.ApproverId);
+                    approver = _mapper.Map<UserBusinessResponse>(approverDataResponse);
+
+                    approverDict.Add(approver.Id, approver);
+                }
+            }
+
+            vacationWithUserBusinessResponses.Add(new VacationWithUserBusinessResponse()
+            {
+                Vacation = vacation,
+                User = user,
+                Aprover = approver
+            });
+        }
+
+        return vacationWithUserBusinessResponses;
     }
 
     public async Task<VacationInfoBusinessResponse> GetVacationInfoByUserIdAsync(Guid userId)
     {
         var curUser = await _userService.GetCurrentUserFromClaimsAsync();
-        if (curUser is null)
-        {
-            throw new ExecutionError("User not found")
-            {
-                Code = GraphQLCustomErrorCodesEnum.USER_NOT_FOUND.ToString()
-            };
-        }
-        else if (!PermissionHelper.HasPermit(curUser.Permissions, PermissionsEnum.GetVacation.ToString()))
+        if (!PermissionHelper.HasPermit(curUser.Permissions, PermissionsEnum.GetVacation.ToString()))
         {
             if (curUser.Id != userId)
             {
@@ -104,6 +155,7 @@ public class VacationService : IVacationService
 
     public async Task<VacationBusinessResponse> CreateVacationAsync(VacationBusinessRequest vacationBusinessRequest)
     {
+        //also need to check if user is not in vacation now / not on sick leave / not fired
         if (DateTime.Compare(vacationBusinessRequest.Start, vacationBusinessRequest.End) > 0)
         {
             throw new ExecutionError("Invalid dates input")
@@ -120,6 +172,7 @@ public class VacationService : IVacationService
 
     public async Task ApproverUpdateVacationAsync(VacationApproveBusinessRequest vacationApproveBusinessRequest)
     {
+        //if vacation has not started yet, approver can update it again
         var vacationApproveDataRequest = _mapper.Map<VacationApproveDataRequest>(vacationApproveBusinessRequest);
         await _vacationRepository.ApproverUpdateVacationAsync(vacationApproveDataRequest);
 
@@ -128,6 +181,8 @@ public class VacationService : IVacationService
             var vacationDataResponse = await _vacationRepository.GetVacationByIdAsync(vacationApproveDataRequest.Id);
             var vacationDurationInDays = (vacationDataResponse.End - vacationDataResponse.Start).TotalDays + 1;
             await _vacationInfoRepository.AddDaysSpentAsync(vacationDataResponse.UserId, (int)vacationDurationInDays);
+
+            await _userRepository.SetUserStatusAsync(vacationDataResponse.UserId, "vacation");
         }
     }
 
@@ -136,6 +191,13 @@ public class VacationService : IVacationService
         var vacationDataResponse = await _vacationRepository.GetVacationByIdAsync(id);
         if (vacationDataResponse is not null)
         {
+            if (vacationDataResponse.IsApproved is not null)
+            {
+                throw new ExecutionError("Can not delete approved vacation")
+                {
+                    Code = GraphQLCustomErrorCodesEnum.OPERATION_FAILED.ToString()
+                };
+            }
             await _vacationRepository.DeleteVacationAsync(id);
         }
     }
